added code for smaller modules, and created a mapper for application questions and user profile fields
