--- utils/application_filler.py
@@ -1,1090 +1,0 @@
-import logging
-from typing import Dict, Any, List
-import asyncio
-import google.generativeai as genai
-from flask import current_app, jsonify
-from models.user import User
-from models.job_recommendation import JobRecommendation
-from flask_login import login_required, current_user
-from playwright.async_api import async_playwright, Playwright
-import tempfile
-import os
-import re
-from application_filler.mappers.field_mapper import map_question_to_field
-import random
-import time
-
-logger = logging.getLogger(__name__)
-
-def valid_url(url: str) -> bool:
-    """
-    Validate if the given URL is a valid job URL.
-    
-    Args:
-        url: The URL to validate
-        
-    Returns:
-        bool: True if the URL is valid, False otherwise
-    """
-    # Example of simple URL validation
-    regex = r"^(https?://)?([a-z0-9-]+\.)+[a-z]{2,6}(/.*)?$"
-    if re.match(regex, url):
-        return True
-    else:
-        logger.warning(f"Invalid URL detected: {url}")
-        return True # Return True for testing purposes
-
-async def extract_application_questions_async(job_id: str, page=None) -> List[Dict[str, Any]]:
-    questions = []
-    launched_browser = False
-
-    if page is None:
-        launched_browser = True
-        from playwright.async_api import async_playwright
-        async with async_playwright() as p:
-            try:
-                # Launch browser with more stable parameters
-                browser = await p.firefox.launch(
-                    headless=False,  # Use headed mode to avoid crashes
-                    args=[
-                        "--no-sandbox",
-                        "--disable-gpu",
-                        "--disable-software-rasterizer",
-                        "--disable-dev-shm-usage",
-                        "--disable-setuid-sandbox"
-                    ]
-                )
-                context = await browser.new_context()
-                page = await context.new_page()
-                # Navigate to the application page
-                await page.goto(job_id, timeout=60000)
-            except Exception as e:
-                logger.error(f"Error launching browser in extract_application_questions_async: {str(e)}")
-                return []
-    else:
-        # Use the provided page and navigate to the URL
-        try:
-            await page.goto(job_id, timeout=60000)
-        except Exception as e:
-            logger.error(f"Error navigating to URL in extract_application_questions_async: {str(e)}")
-            return []
-
-    try:
-        # Wait for form to load
-        logger.info("Waiting for form to load...")
-        await page.wait_for_selector("form", timeout=30000)
-        logger.info("Form loaded, proceeding to extract questions...")
-        
-        # Find all label elements within the form
-        label_elements = await page.query_selector_all("form label")
-        for label_element in label_elements:
-            question_text = await label_element.inner_text()
-            question_text = question_text.strip()
-            
-            # Exclude standard fields
-            if question_text.lower() in ["name", "email", "phone", "resume", "linkedin"]:
-                continue
-            
-            logger.debug(f"Detected question label: {question_text}")
-            
-            # Default question type
-            question_type = "text"
-            questions.append({
-                "text": question_text,
-                "type": question_type
-            })
-    except Exception as e:
-        logger.warning(f"Error finding form elements: {str(e)}")
-        # Return default questions if extraction fails
-        questions = [
-            {"text": "What is your greatest strength?", "type": "text"},
-            {"text": "Why do you want to work here?", "type": "text"},
-        ]
-        logger.info(f"Using default questions: {questions}")
-
-    # If we launched our own browser instance, close it
-    if launched_browser:
-        try:
-            await page.context.browser.close()
-        except Exception as e:
-            logger.warning(f"Error closing browser: {str(e)}")
-
-    logger.info(f"Extracted questions: {questions}")
-    return questions
-
-def extract_application_questions(job_id: str) -> List[Dict[str, Any]]:
-    """
-    Synchronous wrapper for the async extract_application_questions_async function
-    """
-    with current_app.app_context():
-        return asyncio.run(extract_application_questions_async(job_id))
-
-def setup_gemini():
-    """Configure the Gemini API"""
-    api_key = os.environ.get('GEMINI_API_KEY', '')
-    if not api_key:
-        # Try to get from Flask config instead
-        api_key = current_app.config.get('GEMINI_API_KEY', '')
-    
-    if not api_key:
-        # For testing/development, use a fallback mode
-        logger.warning("GEMINI_API_KEY is not set in environment or config. Using mock mode.")
-        
-        # Create a mock configuration to avoid breaking tests/demos
-        class MockGenAI:
-            def configure(*args, **kwargs):
-                pass
-        
-        # Monkey patch the genai module with mock functionality for testing
-        genai.configure = MockGenAI.configure
-        
-        # Return early instead of raising an exception
-        return
-    
-    logger.info(f"Configuring Gemini with API key: {api_key[:5]}...{api_key[-5:] if len(api_key) > 10 else ''}")
-    genai.configure(api_key=api_key)
-
-def generate_application_responses(job_id: str, user: User) -> Dict[str, Any]:
-    """
-    Generate responses to job application questions using Gemini AI
-    
-    Args:
-        job_id: The job ID or URL
-        user: User object with profile information
-        
-    Returns:
-        Dictionary of question responses
-    """
-    # Setup Gemini API
-    setup_gemini()
-    model = genai.GenerativeModel(current_app.config.get('GEMINI_MODEL', 'gemini-pro'))
-    
-    # Extract questions from job application
-    questions = extract_application_questions(job_id)
-    
-    # Prepare dummy responses
-    responses = {question["text"]: "Dummy response" for question in questions}
-    
-    return responses
-
-class ApplicationFiller:
-    def __init__(self, user: User, job_url: str):
-        self.user = user
-        self.job_url = job_url
-        
-        # Initialize user_data from user object for response mapping
-        self.user_data = {
-            "name": self.user.name,
-            "email": self.user.email,
-            "professional_summary": self.user.professional_summary,
-            "skills": self.user.skills,
-            "experience": self.user.experience,
-            "career_goals": self.user.career_goals,
-            "biggest_achievement": self.user.biggest_achievement,
-            "work_style": self.user.work_style,
-            "industry_attraction": self.user.industry_attraction,
-            "willing_to_relocate": self.user.willing_to_relocate,
-            "authorization_status": self.user.authorization_status,
-            "available_start_date": self.user.available_start_date,
-            "portfolio_links": self.user.linkedin_url,
-            "certifications": self.user._certifications if hasattr(self.user, '_certifications') else None,
-            "languages": self.user._languages if hasattr(self.user, '_languages') else None,
-        }
-        
-        # Setup Gemini for response generation
-        setup_gemini()
-        self.model = genai.GenerativeModel(current_app.config.get('GEMINI_MODEL', 'gemini-pro'))
-        
-        # Track failures for retry logic
-        self.failed_fields = []
-        self.max_retries = 3
-
-    async def parse_application_page(self, page) -> list:
-        """
-        Parse the job application page and extract all questions.
-
-        Args:
-            page: The Playwright page object
-
-        Returns:
-            A list of extracted questions
-        """
-        questions = []
-        
-        # Take a screenshot before parsing starts
-        await save_full_page_screenshot(page, "before_parsing_page")
-        
-        try:
-            # Look for form fields directly rather than just labels
-            # This includes inputs, textareas, selects and their associated labels
-            logger.info("Searching for form input fields...")
-            
-            # First, look for all input fields, textareas and select dropdowns
-            form_fields = await page.query_selector_all('input:not([type="hidden"]):not([type="submit"]):not([type="button"]), textarea, select')
-            
-            logger.info(f"Found {len(form_fields)} potential form fields")
-            
-            for field in form_fields:
-                try:
-                    field_type = await field.get_attribute('type') or 'text'
-                    field_name = await field.get_attribute('name') or ''
-                    field_id = await field.get_attribute('id') or ''
-                    field_placeholder = await field.get_attribute('placeholder') or ''
-                    
-                    # Skip login/email/password/name/file fields - these are handled separately
-                    if any(keyword in field_name.lower() or keyword in field_type.lower() or keyword in field_id.lower() 
-                           for keyword in ['email', 'name', 'first', 'last', 'phone', 'resume', 'file', 'password', 'login', 'submit', 'button']):
-                        logger.debug(f"Skipping field: {field_name or field_id} (standard field)")
-                        continue
-                    
-                    # Try to find associated label
-                    question_text = ""
-                    
-                    # First check if we can find a label with a 'for' attribute matching this field's ID
-                    if field_id:
-                        label = await page.query_selector(f'label[for="{field_id}"]')
-                        if label:
-                            question_text = await label.inner_text()
-                    
-                    # If no label was found, try looking for enclosing labels or nearby text
-                    if not question_text and field_name:
-                        # Try to find text elements near this field
-                        label_candidates = await page.query_selector_all(f'label, div, p, span')
-                        for candidate in label_candidates:
-                            candidate_text = await candidate.inner_text()
-                            if field_name.lower() in candidate_text.lower() or field_name.replace('_', ' ').lower() in candidate_text.lower():
-                                question_text = candidate_text
-                                break
-                    
-                    # Use placeholder text if no label was found
-                    if not question_text and field_placeholder:
-                        question_text = field_placeholder
-                        
-                    # Use field name as fallback if still no question text
-                    if not question_text:
-                        question_text = field_name.replace('_', ' ').capitalize()
-                    
-                    # Clean up the question text
-                    question_text = question_text.strip()
-                    if question_text.endswith(':'):
-                        question_text = question_text[:-1]
-                        
-                    if question_text:
-                        logger.info(f"Found form field: '{question_text}' (Type: {field_type}, ID: {field_id or 'none'}, Name: {field_name or 'none'})")
-                        questions.append({
-                            "text": question_text,
-                            "type": field_type,
-                            "element_id": field_id,
-                            "name": field_name,
-                            "field": field
-                        })
-                except Exception as field_error:
-                    logger.error(f"Error processing field: {str(field_error)}")
-        except Exception as e:
-            logger.error(f"Error parsing application page: {str(e)}")
-            
-        # Only use default questions if we're testing with a dummy URL
-        if not questions and ('sample_job_id' in self.job_url or 'example.com' in self.job_url):
-            logger.warning("No questions found and using test URL, using default questions")
-            questions = [
-                {"text": "What is your greatest strength?", "type": "text"},
-                {"text": "Why do you want to work here?", "type": "text"},
-            ]
-        elif not questions:
-            # For real applications, just log that no fields were found
-            logger.warning("No fillable form fields were detected on this page")
-            
-        return questions
-
-    async def fill_application_field(self, page, question: dict, user_response: str, retry_count=0):
-        """
-        Fill out a single application field based on the extracted question and user profile data.
-        """
-        question_text = question["text"]
-        question_type = question.get("type", "text")
-        element_id = question.get("element_id")
-        
-        if retry_count > self.max_retries:
-            logger.error(f"Max retries exceeded for question '{question_text}'")
-            self.failed_fields.append(question_text)
-            return False
-            
-        try:
-            # Debug screenshot before attempting to fill field
-            debug_path = f"/tmp/before_fill_{int(time.time())}.png"
-            await page.screenshot(path=debug_path)
-            logger.info(f"Debug screenshot saved to {debug_path}")
-            
-            # First try by id if available
-            if element_id:
-                selector = f"#{element_id}"
-                logger.info(f"Using ID selector: {selector}")
-                
-                # Direct fill approach for ID-based elements
-                try:
-                    await page.fill(selector, user_response)
-                    logger.info(f"Filled field with ID '{element_id}' using direct fill")
-                    return True
-                except Exception as e:
-                    logger.warning(f"Direct fill failed, trying alternative methods: {str(e)}")
-            else:
-                # Try multiple selector strategies - improved for better detection
-                selectors = [
-                    f'input[placeholder*="{question_text}"], textarea[placeholder*="{question_text}"]',
-                    f'label:has-text("{question_text}") + input, label:has-text("{question_text}") + textarea',
-                    f'div:has-text("{question_text}") input, div:has-text("{question_text}") textarea',
-                    f'input[name*="{question_text.lower().replace(" ", "_")}"], textarea[name*="{question_text.lower().replace(" ", "_")}"]',
-                    'form input[type="text"]:visible, form textarea:visible'  # Fallback to any visible text input
-                ]
-                
-                # Try each selector until one works
-                found_element = None
-                working_selector = None
-                
-                for selector in selectors:
-                    logger.info(f"Trying selector: {selector}")
-                    try:
-                        found_element = await page.query_selector(selector)
-                        if found_element:
-                            logger.info(f"Found element with selector: {selector}")
-                            working_selector = selector
-                            break
-                    except Exception as e:
-                        logger.debug(f"Selector {selector} failed: {str(e)}")
-                
-                if found_element and working_selector:
-                    # Try direct fill first
-                    try:
-                        await found_element.fill(user_response)
-                        logger.info(f"Filled using element.fill() method")
-                        return True
-                    except Exception as e:
-                        logger.warning(f"Element.fill() failed: {str(e)}, trying JavaScript")
-                        selector = working_selector
-                else:
-                    # If all selectors fail, try a more generic approach
-                    logger.warning(f"No element found for question '{question_text}'. Using fallback approach.")
-                    try:
-                        # Just find any input or textarea
-                        selector = 'input[type="text"], textarea'
-                        found_element = await page.query_selector(selector)
-                        if not found_element:
-                            raise Exception("No input elements found")
-                    except Exception as e:
-                        logger.error(f"Failed to find any input element: {str(e)}")
-                        self.failed_fields.append(question_text)
-                        return False
-            
-            if not user_response:
-                logger.warning(f"No response available for question '{question_text}'.")
-                return False
-            
-            # Use JavaScript evaluation with correct syntax (only 2 arguments: js_code and selector+value object)
-            await page.evaluate("""
-                (args) => {
-                    const elements = document.querySelectorAll(args.selector);
-                    if (elements.length > 0) {
-                        const element = elements[0];
-                        element.value = args.value;
-                        element.dispatchEvent(new Event('input', { bubbles: true }));
-                        element.dispatchEvent(new Event('change', { bubbles: true }));
-                        return true;
-                    }
-                    return false;
-                }
-            """, {"selector": selector, "value": user_response})
-            
-            logger.info(f"Filled field for '{question_text}' with response starting: {user_response[:30]}...")
-            
-            # Add a randomized delay to appear more human-like
-            delay_time = random.uniform(1.5, 3.5)
-            await asyncio.sleep(delay_time)
-            return True
-            
-        except Exception as e:
-            logger.error(f"Error filling field for question '{question_text}': {str(e)}")
-            
-            # Take screenshot of the issue
-            error_screenshot = f"/tmp/error_fill_{int(time.time())}.png"
-            try:
-                await page.screenshot(path=error_screenshot)
-                logger.info(f"Error screenshot saved to {error_screenshot}")
-            except Exception as screenshot_e:
-                logger.warning(f"Failed to take error screenshot: {str(screenshot_e)}")
-            
-            # Retry with increased delay
-            if retry_count < self.max_retries:
-                logger.info(f"Retrying field '{question_text}' (Attempt {retry_count + 1}/{self.max_retries})")
-                await asyncio.sleep(2 + retry_count)  # Increasing delay with each retry
-                return await self.fill_application_field(page, question, user_response, retry_count + 1)
-            else:
-                self.failed_fields.append(question_text)
-                return False
-
-    def generate_response_based_on_question(self, question_text: str) -> str:
-        """
-        Generate an appropriate response for an application question based on user profile.
-        
-        Args:
-            question_text: The question to answer
-            
-        Returns:
-            String response to the question
-        """
-        # First try to map question to a field in user data
-        mapped_field = map_question_to_field(question_text)
-        
-        # If we have user data for this question type, use it
-        if mapped_field in self.user_data and self.user_data[mapped_field]:
-            logger.info(f"Using profile data field '{mapped_field}' for question: {question_text}")
-            return str(self.user_data[mapped_field])
-        
-        # If no direct mapping or empty data, use AI to generate response
-        try:
-            # Create prompt for Gemini
-            prompt = f"""
-            As {self.user.name}, please provide a professional response to this job application question:
-            
-            Question: "{question_text}"
-            
-            Use this profile information to personalize your answer:
-            - Professional summary: {self.user.professional_summary or ''}
-            - Skills: {self.user.skills or ''}
-            - Experience: {self.user.experience or ''}
-            - Career goals: {self.user.career_goals or ''}
-            - Work style: {self.user.work_style or ''}
-            
-            Write in first person as {self.user.name}.
-            Keep your answer professional, concise (2-3 sentences) and specifically tailored to the question.
-            """
-            
-            # Check if we're in mock mode (no API key available)
-            api_key = current_app.config.get('GEMINI_API_KEY', os.environ.get('GEMINI_API_KEY', ''))
-            
-            if not api_key:
-                logger.warning("Using mock response generation for: " + question_text)
-                
-                # Generate deterministic mock responses based on question type
-                if "strength" in question_text.lower() or "skill" in question_text.lower():
-                    return f"My greatest strength is my ability to {self.user_data.get('skills', 'solve complex problems')}. I've demonstrated this through my past work where I consistently delivered results."
-                    
-                elif "weakness" in question_text.lower():
-                    return "I'm always working to improve my time management skills. I've implemented structured planning systems that have significantly improved my productivity."
-                    
-                elif "experience" in question_text.lower() or "background" in question_text.lower():
-                    return f"I have experience in {self.user_data.get('experience', 'software development and project management')}. This has given me a strong foundation in delivering quality work."
-                    
-                elif "why" in question_text.lower() and "work" in question_text.lower():
-                    return "I'm excited about this opportunity because it aligns with my professional goals. I believe my skills would be a great match for this position."
-                    
-                else:
-                    return "I'm excited about this opportunity and believe my skills and experience make me a strong candidate. I look forward to potentially joining your team."
-            else:
-                # Use the real Gemini API
-                response = self.model.generate_content(prompt)
-                answer = response.text.strip()
-                logger.info(f"Generated AI response for question: {question_text[:30]}...")
-                return answer
-            
-        except Exception as e:
-            logger.error(f"Error generating response with AI: {str(e)}")
-            return f"I believe I am well-qualified for this position with my background in {self.user_data.get('skills', 'professional skills')} and I'm excited about this opportunity."
-
-    async def handle_resume_upload(self, page):
-        """
-        Handle the upload of the resume if a file input is present.
-
-        Args:
-            page: The Playwright page object
-        """
-        if self.user.resume_file_path and os.path.exists(self.user.resume_file_path):
-            try:
-                logger.info(f"Uploading resume from: {self.user.resume_file_path}")
-                
-                # Try different selectors for file upload inputs
-                file_input_selectors = [
-                    'input[type="file"]',
-                    'input[name="resume"]',
-                    'input[accept=".pdf,.doc,.docx"]',
-                    'label:has-text("Resume") input[type="file"]',
-                    'label:has-text("CV") input[type="file"]'
-                ]
-                
-                for selector in file_input_selectors:
-                    file_input = await page.query_selector(selector)
-                    if file_input:
-                        await file_input.set_input_files(self.user.resume_file_path)
-                        logger.info(f"Resume uploaded successfully using selector: {selector}")
-                        await asyncio.sleep(2)  # Wait after upload
-                        return True
-                
-                logger.warning("Resume file input not found after trying multiple selectors.")
-                return False
-                
-            except Exception as e:
-                logger.error(f"Error uploading resume: {str(e)}")
-                return False
-        else:
-            logger.warning("No resume file path or file doesn't exist.")
-            return False
-
-    async def fill_application_form(self, page):
-        """
-        Fill the job application form using user profile data.
-
-        Args:
-            page: The Playwright page object
-            
-        Returns:
-            Dictionary with success status and list of any failed fields
-        """
-        logger.info(f"Starting form fill for job URL: {self.job_url} using user: {self.user.email}")
-        
-        # Reset failed fields tracking
-        self.failed_fields = []
-        
-        # First extract questions from the page
-        questions = await self.parse_application_page(page)
-        logger.info(f"Found {len(questions)} questions to fill")
-        
-        # Fill each question's field
-        for i, question in enumerate(questions):
-            logger.info(f"Processing question {i+1}/{len(questions)}: {question['text']}")
-            
-            # Generate response for this question
-            response = self.generate_response_based_on_question(question["text"])
-            
-            # Fill the field with the generated response
-            await self.fill_application_field(page, question, response)
-        
-        # Return success status and any failures
-        success = len(self.failed_fields) == 0
-        return {
-            "success": success,
-            "total_questions": len(questions),
-            "failed_fields": self.failed_fields
-        }
-
-    async def fill_application(self):
-        """
-        Main function that handles browser creation, parsing, filling, and submission of the application.
-        
-        Returns:
-            Dictionary with application results and status
-        """
-        result = {
-            "success": False,
-            "message": "",
-            "url": self.job_url,
-            "user": self.user.email,
-            "form_completed": False,
-            "resume_uploaded": False,
-            "failed_fields": []
-        }
-        
-        logger.info(f"Starting application process for {self.job_url}")
-        
-        try:
-            async with async_playwright() as p:
-                # Launch browser with more stable parameters
-                browser = await p.firefox.launch(
-                    headless=False,  # Keep browser visible
-                    slow_mo=100,    # Slow down automation for stability
-                    args=[
-                        "--no-sandbox",
-                        "--disable-gpu",
-                        "--disable-dev-shm-usage",
-                    ]
-                )
-                
-                # Create a context with viewport size and user agent
-                context = await browser.new_context(
-                    viewport={"width": 1280, "height": 720},
-                    user_agent="Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/92.0.4515.131 Safari/537.36"
-                )
-                
-                # Open new page and navigate
-                page = await context.new_page()
-                await page.goto(self.job_url, timeout=60000)
-                
-                # Wait for the page to load
-                try:
-                    await page.wait_for_load_state("networkidle", timeout=30000)
-                    logger.info("Page loaded successfully")
-                except Exception as e:
-                    logger.warning(f"Page loading timed out, continuing anyway: {str(e)}")
-                
-                # Take full page screenshot and save HTML for analysis
-                await save_full_page_screenshot(page, "initial_page")
-                
-                # Add manual delay to ensure page is fully loaded
-                await asyncio.sleep(5)
-                
-                # Check if we need to click an "Apply" button or similar first
-                apply_button_found = await self._check_and_click_apply_button(page)
-                if apply_button_found:
-                    # Wait for navigation after clicking apply button
-                    try:
-                        await page.wait_for_load_state("networkidle", timeout=15000)
-                        # Save screenshot of the application form page
-                        await save_full_page_screenshot(page, "after_apply_button")
-                    except Exception as e:
-                        logger.warning(f"Page loading after apply button timed out: {str(e)}")
-                
-                # Look for common application form patterns and handle them
-                form_found = await self._detect_and_handle_form_type(page)
-                if not form_found:
-                    result["message"] = "Could not detect application form on the page"
-                    logger.warning(result["message"])
-                    
-                    # Save final screenshot
-                    await save_full_page_screenshot(page, "final_state")
-                    
-                    # Keep browser open longer if debug mode is enabled
-                    if os.environ.get('DEBUG_MODE') == '1':
-                        logger.info("DEBUG_MODE enabled - keeping browser open for 60 seconds")
-                        await asyncio.sleep(60)
-                    return result
-                
-                # Step 1: Upload resume FIRST before filling the form
-                # This should trigger auto-population of fields in many application systems
-                resume_uploaded = await self.prioritize_resume_upload(page)
-                result["resume_uploaded"] = resume_uploaded
-                
-                if resume_uploaded:
-                    # Wait for potential auto-fill to happen
-                    logger.info("Resume uploaded successfully, waiting for potential autofill...")
-                    await asyncio.sleep(5)
-                    await save_full_page_screenshot(page, "after_resume_upload_autofill")
-                
-                # Step 2: Fill in the basic identifier fields (name, email, etc.)
-                # Look for common basic fields that may be required
-                basic_fields = [
-                    {'selector': 'input[name*="name"], input[placeholder*="name"]', 'value': self.user.name},
-                    {'selector': 'input[type="email"], input[name*="email"], input[placeholder*="email"]', 'value': self.user.email},
-                    {'selector': 'input[name*="phone"], input[placeholder*="phone"], input[type="tel"]', 'value': self.user.phone if hasattr(self.user, 'phone') else ''},
-                ]
-                
-                for field in basic_fields:
-                    if field['value']:  # Only try to fill if we have a value
-                        try:
-                            elements = await page.query_selector_all(field['selector'])
-                            for elem in elements:
-                                is_visible = await elem.is_visible()
-                                if is_visible:
-                                    await elem.fill(field['value'])
-                                    logger.info(f"Filled basic field: {field['selector']} with value: {field['value']}")
-                                    break
-                        except Exception as e:
-                            logger.debug(f"Error filling basic field {field['selector']}: {str(e)}")
-                
-                # Wait a bit after filling basic fields
-                await asyncio.sleep(2)
-                
-                # Step 3: Now fill the rest of the form
-                form_result = await self.fill_application_form(page)
-                result["form_completed"] = form_result["success"]
-                result["failed_fields"] = form_result["failed_fields"]
-                
-                # Check if there's a submit button and click it
-                submit_clicked = await self._find_and_click_submit_button(page)
-                if submit_clicked:
-                    logger.info("Submit button clicked successfully")
-                    
-                    # Wait a bit for submission to complete
-                    await asyncio.sleep(5)
-                    
-                    # Save screenshot of confirmation page
-                    await save_full_page_screenshot(page, "confirmation_page")
-                
-                # Set overall success based on form completion and resume upload
-                result["success"] = result["form_completed"]
-                
-                # Keep browser open for manual inspection if requested (dev mode)
-                if os.environ.get('DEBUG_MODE') == '1':
-                    logger.info("DEBUG_MODE enabled - keeping browser open for 30 seconds")
-                    await asyncio.sleep(30)
-                
-                # Set final message
-                if result["success"]:
-                    result["message"] = "Application form filled successfully."
-                    if submit_clicked:
-                        result["message"] += " Submit button was clicked."
-                    logger.info("Application process completed successfully")
-                else:
-                    result["message"] = f"Application partially completed with {len(form_result['failed_fields'])} failed fields."
-                    logger.warning(f"Application process partially completed: {result['message']}")
-                
-                # Close browser
-                await browser.close()
-                
-        except Exception as e:
-            error_message = f"Error during application process: {str(e)}"
-            result["message"] = error_message
-            result["success"] = False
-            logger.error(error_message)
-            
-        return result
-
-    async def _check_and_click_apply_button(self, page):
-        """
-        Look for and click an "Apply" button if one exists.
-        
-        Returns:
-            bool: True if an apply button was found and clicked
-        """
-        # Common selectors for apply buttons
-        apply_button_selectors = [
-            'button:has-text("Apply")', 
-            'button:has-text("Apply Now")',
-            'a:has-text("Apply")',
-            'a:has-text("Apply Now")',
-            'a.apply-button',
-            'button.apply-button',
-            '[data-automation="apply-button"]',
-            '[aria-label*="apply"]'
-        ]
-        
-        for selector in apply_button_selectors:
-            try:
-                logger.info(f"Looking for apply button with selector: {selector}")
-                button = await page.query_selector(selector)
-                if button:
-                    # Check if button is visible
-                    is_visible = await button.is_visible()
-                    if is_visible:
-                        logger.info(f"Found apply button with selector: {selector}")
-                        await button.click()
-                        logger.info("Clicked apply button")
-                        return True
-            except Exception as e:
-                logger.debug(f"Error checking apply button selector '{selector}': {str(e)}")
-        
-        logger.info("No apply button found or could not be clicked")
-        return False
-    
-    async def _detect_and_handle_form_type(self, page):
-        """
-        Detect what type of application form we're dealing with and handle it appropriately.
-        
-        Returns:
-            bool: True if a form was detected and handled
-        """
-        # Take a screenshot before detection
-        await save_full_page_screenshot(page, "before_form_detection")
-        
-        # Check for different types of application forms
-        form_types = [
-            self._check_for_standard_form,
-            self._check_for_greenhouse_form,
-            self._check_for_lever_form,
-            self._check_for_workday_form,
-            self._check_for_ashby_form,
-            self._check_for_indeed_form,
-            self._check_for_linkedin_form
-        ]
-        
-        for check_function in form_types:
-            try:
-                logger.info(f"Trying form detection with {check_function.__name__}")
-                form_detected = await check_function(page)
-                if form_detected:
-                    logger.info(f"Form detected with {check_function.__name__}")
-                    return True
-            except Exception as e:
-                logger.error(f"Error in {check_function.__name__}: {str(e)}")
-        
-        logger.warning("No form type detected")
-        return False
-    
-    async def _check_for_standard_form(self, page):
-        """Check for a standard HTML form"""
-        form = await page.query_selector('form')
-        if (form):
-            logger.info("Standard form detected")
-            return True
-        return False
-    
-    async def _check_for_greenhouse_form(self, page):
-        """Check for Greenhouse ATS form"""
-        if 'greenhouse' in await page.content():
-            logger.info("Greenhouse form detected")
-            # Handle Greenhouse specific logic here
-            return True
-        return False
-    
-    async def _check_for_lever_form(self, page):
-        """Check for Lever ATS form"""
-        if 'lever' in await page.content():
-            logger.info("Lever form detected")
-            # Handle Lever specific logic here
-            return True
-        return False
-    
-    async def _check_for_workday_form(self, page):
-        """Check for Workday ATS form"""
-        if 'workday' in await page.content():
-            logger.info("Workday form detected")
-            # Handle Workday specific logic here
-            return True
-        return False
-
-    async def _check_for_ashby_form(self, page):
-        """Check for Ashby ATS form"""
-        if 'ashbyhq' in self.job_url or 'ashby' in await page.content():
-            logger.info("Ashby form detected")
-            
-            # Look for and click the first step in the application
-            try:
-                # Check if there's a "start application" or similar button
-                start_buttons = [
-                    'button:has-text("Start")',
-                    'button:has-text("Begin")',
-                    'button:has-text("Start Application")',
-                    'a:has-text("Start Application")'
-                ]
-                
-                for selector in start_buttons:
-                    button = await page.query_selector(selector)
-                    if button and await button.is_visible():
-                        await button.click()
-                        logger.info(f"Clicked start button with selector: {selector}")
-                        await asyncio.sleep(2)
-                        await save_full_page_screenshot(page, "after_ashby_start_button")
-                        break
-                
-                # Look for name input fields - Ashby typically has these at the beginning
-                name_fields = await page.query_selector_all('input[name*="name"], input[placeholder*="name"]')
-                if name_fields:
-                    for field in name_fields:
-                        field_name = await field.get_attribute('name') or await field.get_attribute('placeholder') or "unknown"
-                        if "first" in field_name.lower():
-                            await field.fill(self.user.name.split()[0])
-                        elif "last" in field_name.lower():
-                            if len(self.user.name.split()) > 1:
-                                await field.fill(self.user.name.split()[1])
-                        else:
-                            await field.fill(self.user.name)
-                
-                # Look for email input
-                email_field = await page.query_selector('input[type="email"], input[name*="email"], input[placeholder*="email"]')
-                if email_field:
-                    await email_field.fill(self.user.email)
-                
-                # Look for "Next" button to proceed to questions
-                next_buttons = [
-                    'button:has-text("Next")',
-                    'button:has-text("Continue")',
-                    'button[type="submit"]',
-                    'button.next-button'
-                ]
-                
-                for selector in next_buttons:
-                    button = await page.query_selector(selector)
-                    if button and await button.is_visible():
-                        await button.click()
-                        logger.info(f"Clicked next button with selector: {selector}")
-                        await asyncio.sleep(3)
-                        await save_full_page_screenshot(page, "after_ashby_next_button")
-                        break
-            except Exception as e:
-                logger.error(f"Error handling Ashby form: {str(e)}")
-            
-            return True
-        return False
-    
-    async def _check_for_indeed_form(self, page):
-        """Check for Indeed application form"""
-        if 'indeed' in self.job_url or 'indeed' in await page.content():
-            logger.info("Indeed form detected")
-            
-            # Handle Indeed specific logic here
-            try:
-                # Look for "Apply now" button
-                apply_buttons = await page.query_selector_all('button:has-text("Apply now"), a:has-text("Apply now")')
-                if apply_buttons:
-                    await apply_buttons[0].click()
-                    logger.info("Clicked 'Apply now' button on Indeed")
-                    await asyncio.sleep(2)
-            except Exception as e:
-                logger.error(f"Error with Indeed apply button: {str(e)}")
-                
-            return True
-        return False
-    
-    async def _check_for_linkedin_form(self, page):
-        """Check for LinkedIn EasyApply form"""
-        if 'linkedin.com/jobs' in self.job_url:
-            logger.info("LinkedIn job page detected")
-            
-            try:
-                # Look for "Easy Apply" button
-                easy_apply_buttons = [
-                    'button:has-text("Easy Apply")',
-                    'button:has-text("Apply")',
-                    'button.jobs-apply-button'
-                ]
-                
-                for selector in easy_apply_buttons:
-                    button = await page.query_selector(selector)
-                    if button and await button.is_visible():
-                        await button.click()
-                        logger.info(f"Clicked LinkedIn apply button with selector: {selector}")
-                        await asyncio.sleep(3)
-                        await save_full_page_screenshot(page, "after_linkedin_apply_button")
-                        
-                        # LinkedIn often has a multi-step application form
-                        # Look for next button to navigate through steps
-                        next_buttons = [
-                            'button:has-text("Next")',
-                            'button:has-text("Continue")',
-                            'button[aria-label="Continue to next step"]'
-                        ]
-                        
-                        for next_selector in next_buttons:
-                            next_button = await page.query_selector(next_selector)
-                            if next_button and await next_button.is_visible():
-                                await next_button.click()
-                                logger.info("Clicked LinkedIn next button")
-                                await asyncio.sleep(2)
-                                await save_full_page_screenshot(page, "linkedin_next_step")
-                                break
-                        
-                        return True
-            except Exception as e:
-                logger.error(f"Error with LinkedIn apply button: {str(e)}")
-                
-            return True
-        return False
-    
-    async def _find_and_click_submit_button(self, page):
-        """
-        Find and click the submit button to complete the application.
-        
-        Returns:
-            bool: True if submit button was found and clicked
-        """
-        # Common submit button selectors
-        submit_selectors = [
-            'button[type="submit"]',
-            'button:has-text("Submit")',
-            'button:has-text("Submit Application")',
-            'button:has-text("Apply")',
-            'input[type="submit"]',
-            'button.submit-button',
-            'button:has-text("Send Application")',
-            'button:has-text("Complete Application")'
-        ]
-        
-        # Take a screenshot before looking for submit button
-        await save_full_page_screenshot(page, "before_submit_button")
-        
-        for selector in submit_selectors:
-            try:
-                logger.info(f"Looking for submit button with selector: {selector}")
-                submit_button = await page.query_selector(selector)
-                if submit_button:
-                    is_visible = await submit_button.is_visible()
-                    if is_visible:
-                        logger.info(f"Found submit button with selector: {selector}")
-                        await submit_button.click()
-                        logger.info("Clicked submit button")
-                        return True
-            except Exception as e:
-                logger.debug(f"Error checking submit button selector '{selector}': {str(e)}")
-                
-        logger.warning("No submit button found or could not be clicked")
-        return False
-
-    async def prioritize_resume_upload(self, page):
-        """
-        Upload resume first, before filling any other fields.
-        This is important because some forms auto-populate fields from the resume.
-        
-        Args:
-            page: The Playwright page object
-            
-        Returns:
-            bool: True if resume was uploaded
-        """
-        logger.info("Looking for resume upload field first")
-        
-        # First take a screenshot to see the current page
-        await save_full_page_screenshot(page, "before_resume_upload")
-        
-        # Common selectors for resume upload fields
-        resume_selectors = [
-            'input[type="file"][accept*="pdf"]',
-            'input[type="file"][accept*="doc"]',
-            'input[type="file"]',
-            'label:has-text("Resume") input[type="file"]',
-            'label:has-text("Upload") input[type="file"]',
-            'label:has-text("CV") input[type="file"]',
-            'div:has-text("Upload resume") input[type="file"]',
-            'div:has-text("Upload your resume") input[type="file"]',
-            'div[aria-label*="resume"] input[type="file"]'
-        ]
-        
-        if self.user.resume_file_path and os.path.exists(self.user.resume_file_path):
-            for selector in resume_selectors:
-                try:
-                    logger.info(f"Looking for resume upload with selector: {selector}")
-                    file_input = await page.wait_for_selector(selector, timeout=3000)
-                    if file_input:
-                        logger.info(f"Found resume upload field with selector: {selector}")
-                        await file_input.set_input_files(self.user.resume_file_path)
-                        logger.info(f"Resume uploaded from: {self.user.resume_file_path}")
-                        
-                        # Wait for autofill to potentially occur
-                        await asyncio.sleep(5)
-                        
-                        # Check if there's a "parse resume" or similar button that needs to be clicked
-                        parse_buttons = [
-                            'button:has-text("Parse")',
-                            'button:has-text("Extract")',
-                            'button:has-text("Autofill")',
-                            'a:has-text("Autofill")',
-                            'button.parse-resume'
-                        ]
-                        
-                        for btn_selector in parse_buttons:
-                            parse_button = await page.query_selector(btn_selector)
-                            if parse_button and await parse_button.is_visible():
-                                await parse_button.click()
-                                logger.info(f"Clicked parse button: {btn_selector}")
-                                await asyncio.sleep(3)  # Wait for parsing to complete
-                                break
-                        
-                        # Take screenshot after upload
-                        await save_full_page_screenshot(page, "after_resume_upload")
-                        return True
-                except Exception as e:
-                    logger.debug(f"Error with resume selector {selector}: {str(e)}")
-            
-            logger.warning("Could not find resume upload field")
-            return False
-        else:
-            logger.warning("No resume file path found or file does not exist")
-            return False
-
-async def save_full_page_screenshot(page, name_prefix="full_page"):
-    """Take a full page screenshot with useful debug information."""
-    try:
-        # Create a more detailed filename with timestamp
-        filename = f"/tmp/{name_prefix}_{int(time.time())}.png"
-        
-        # Take a full page screenshot
-        await page.screenshot(path=filename, full_page=True)
-        
-        # Also capture page HTML for analysis
-        html_filename = f"/tmp/{name_prefix}_html_{int(time.time())}.html"
-        html_content = await page.content()
-        with open(html_filename, 'w', encoding='utf-8') as f:
-            f.write(html_content)
-        
-        logger.info(f"Full page screenshot saved to {filename}")
-        logger.info(f"HTML content saved to {html_filename}")
-        return filename
-    except Exception as e:
-        logger.error(f"Failed to take full page screenshot: {str(e)}")
-        return None
